Improve Loading speeds
The app now caches the TISS cookies, so it doesn't have to go through
the whole (and very slow) login-process everytime.

It saves the TISS cookie for 30min and deletes it if you edit your
username/password in the settings.

This means that the first search after you set your password or after
30mins will take notably longer. This can (and will) be improved by
going through the login process before the user even clicks on the
searchbutton.