Switch to English
While some API fields are always German and the languages are now more
consistent than before. I still think that this is the right way to
go as there are students/staff that cannot speak German and English is
the de facto norm in the academic world anyway.

Of course the right way would be to support English and German, but
internationalization seams way too hard (and too little fun) for a
side project. Maybe I will reconsider in the future if more people start
using the app.